Unify process package into data
