+++ cyclops/data/constants.py
