--- tests/cyclops/process/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for processors API package."""
